requirements.txt
python-dotenv==1.0.1
setuptools==72.2.0
pyspark==3.2.0
Faker==28.0.0